finish matrix transpose

+++ 05_Matrix_manipulation/Source_code/03matrices_additionv2_P/03matrices_additionv2.cu
@@ -5,7 +5,7 @@
 
 #define N 500
 
-__device__ int calcIndex(int x, int y) {
+__device__ size_t calcIndex(size_t x, size_t y) {
     return y * N + x;
 }
